requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-environ==0.11.2
django-jazzmin==3.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
multidict==6.0.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
setuptools==71.1.0
sqlparse==0.5.1
twilio==9.2.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
